Added a string helper to replace accent. Fixed some search bugs
